fix: Remove circular context exports from barrel file for more consistent HMR
